test(trial): Update tests to reflect payment optionality as always-true.
